Working Channel model
Max_Temp_Single_Channels: outputs maximum temp in channel given power,
height, length of run, header pressures and temperature of entering
liquid. For reverse flow the incoming vapor is currently set as the same
as the saturation temperature at the pressure in the channel.

Height_vs_Power: this determines the temperatures and flowrate for a
range of different temperatures and heights.

Qzirc_Steam; Initial attempt to account for zirc-steam reaction at high
temperatures. Has not been implemented yet.